Use a more universal sort function
Summary: Abort on missing tag

Test Plan:
Removed the current lib, then ran the installer. It fails, as
expected, because we don't have a stable version. Then ran the installer
with install rc, which found the latest -rc tag and installed it.

Also, the sort function was tested on my macbook air and linux terminal
with the git tags in project13, and adding "1.0.1", "1.0.1-dev", and
"1.0.1-RC4"

Reviewers: jpcivile

Reviewed By: jpcivile

Subscribers: jmsotuyo

Differential Revision: http://ph.monits.com/D5118