Fix path to binary, files and lint order
Summary: Actually add the files to the list

Test Plan:
The test procedure was to run it on AndroidAgilefant with changes
in some xml files and java source code. Then to drop the changes to the xml
files and run it again.

Tested it without the android sdk installed, without the $ANDROID_HOME,
with changes, without changes, with the sdk installed and $ANDROID_HOME unset,
and finally with changes (and then without changes) with the sdk installed and
$ANDROID_HOME set.

Reviewers: jpcivile

Reviewed By: jpcivile

Subscribers: jmsotuyo

Differential Revision: http://ph.monits.com/D5154