Add initial stage of the indentation linter
Summary:
The aim of the **indentation linter** will be to check the consistency of the
indentation inside XML and JSON files.

In this first commit, we are modifying the `__phputil_library_map__` (by using
`arc liberator`) to add the new linter.

Also, the submitted linter has the following features:

* It can detect automatically which type of indentation is being used in the
file.
* It can warn about mixed indentation.

To guess which kind of indentation it has, it uses the mechanism `comparing
lines`, which is explained in:

https://medium.com/firefox-developer-tools/detecting-code-indentation-eff3ed0fb56b#.r1qh0v3sv

Future features:

* Parse XML files, and guess what type of indentation is required for a certain
level.
* Procceed with JSON files in a similar fashion.

Test Plan:
Checked by testing arc lint in a sample repository, with XML files
with different kinds of indentation.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Subscribers: pcostesich

Differential Revision: http://ph.monits.com/D13862