avoid full rebuilds for OCLint compile commands generation
Summary: For OCLint we need to generate the compile_commands.json file. This requires running `xcodebuild | xcpretty`. Luckily, we can use `-dry-run` to avoid actually compiling files in this step, but it still causes `xcodebuild` to process the whole project. We now use `touch` to trigger an incremental (dry run'ed) rebuild of only the files to be linted.

Test Plan:
```
for i in 1 2 3 4; do
    echo "lint";
    time '/Users/juanpablocivile/Documents/arc/arcanist/bin/../scripts/arcanist.php' 'lint' 'LibraryComponents//Classes/Core/Models/MLQQuestionStatusFactory.m' 'LibraryComponents//Classes/Core/Models/MLQStockItemSizes.m' 'LibraryComponents//Classes/Core/Services/MLQPendingQuestionsService.m' 'LibraryComponents//Classes/Core/Services/MLQServiceFactory.m' 'LibraryComponents//Classes/Core/ViewControllers/MLQEmptyQuestionsViewController.m' 'LibraryComponents//Classes/Core/ViewControllers/MLQKeyboardViewController.m' 'LibraryComponents//Classes/Core/ViewModel/MLQItemDetailViewModel.m' 'LibraryComponents//Classes/Core/Views/MLQAnswerBubbleCell.m' 'LibraryComponents//Classes/Core/Views/MLQAttachmentCell.m' 'LibraryComponents//Classes/Core/Views/MLQBubbleStyle.m' 'LibraryComponents//Classes/Core/Views/MLQContentView.m' 'LibraryComponents//Classes/Core/Views/MLQEmptyQuestionBackgroundView.m' 'LibraryComponents//Classes/Core/Views/MLQGenericBannerErrorView.m' 'LibraryComponents/Classes/Seller/Views/MLQSellerStockHeaderView.m' > /dev/null;
    echo "unit";
    time arc unit > /dev/null ;
done
```

Linting 14 files, test results oscilate between 3 and 7 seconds improvements, with all results consistently lower.
For a small number of files there's a 3 second improvement, which is compensated by a 3 second increase in unit test running.

Reviewers: jmsotuyo

Reviewed By: jmsotuyo

Differential Revision: http://ph.monits.com/D16537